requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\Python\vojimir\main.py: 1
discord == 1.0.1

# D:\Python\vojimir\database.py: 4
gspread == 3.6.0

# D:\Python\vojimir\database.py: 3
oauth2client == 4.1.3

# D:\Python\vojimir\botFunctions.py: 2
smaz-py3 == 1.0.1

bdbf == 0.1.0
